updated 2022.02.21

+++ source/augmentation/bbox_trick.py
@@ -8,7 +8,7 @@ from src.utils import read_xml, write_xml, read_label_file, get_files, visualize
 if __name__ == "__main__":
     ROOT_DIR = "/data/Datasets/SPC"
     FOLDER = f"{ROOT_DIR}/Cvat"
-    SAVE_DIR = f"{ROOT_DIR}/full-name10"
+    SAVE_DIR = f"{ROOT_DIR}/full-name-test"
     LABEL_DIR = f"{ROOT_DIR}/Labels/labels.txt"
     TARGETS = ["Paris_baguette"]
 
@@ -33,45 +33,39 @@ if __name__ == "__main__":
 
         print(len(images), len(annotations))
         for idx in tqdm(range(len(images))):
-            try:
-                image = images[idx]
-                annot = annotations[idx]
-                filename = image.split('/')[-1].split('.')[0]
-                image = cv2.imread(image)
-                height, width = image.shape[:-1]
-                bboxes, labels = read_xml(annot, classes, format='pascal_voc')
+            image = images[idx]
+            annot = annotations[idx]
+            filename = image.split('/')[-1].split('.')[0]
+            image = cv2.imread(image)
+            height, width = image.shape[:-1]
+            bboxes, labels = read_xml(annot, classes, format='pascal_voc')
 
-                for label in labels:
-                    label_check_list.add(label)
+            for label in labels:
+                label_check_list.add(label)
 
-                transformed = transform(image=image, bboxes=bboxes, labels=labels)
-                t_image, t_bboxes, t_labels = transformed['image'], transformed['bboxes'], transformed['labels']
+            transformed = transform(image=image, bboxes=bboxes, labels=labels)
+            t_image, t_bboxes, t_labels = transformed['image'], transformed['bboxes'], transformed['labels']
 
-                confirm = []
-                for xmin, ymin, xmax, ymax in t_bboxes:
-                    if (xmax - xmin) / (ymax - ymin) > limit_ratio:               
-                        # ymed = (ymax + ymin) // 2
-                        # yhlf = (xmax - xmin) // (2 * limit_ratio)
-                        # ymin = max(0, ymed - yhlf)
-                        # ymax = min(height - 1, ymed + yhlf)
-                        if ymin - GAP > 0 and ymax < height - 1:
-                            ymin -= GAP
-                            ymax += GAP
+            confirm = []
+            for xmin, ymin, xmax, ymax in t_bboxes:
+                if (xmax - xmin) / (ymax - ymin) > limit_ratio:
+                    print(annot)               
+                    if (ymin - GAP) > 0 and (ymax + GAP) < height - 1:
+                        ymin -= GAP
+                        ymax += GAP            
+                        confirm.append((xmin, ymin, xmax, ymax))
 
-                        else:
-                            ymin = 0
-                            ymax = height - 1
-                        
-                    confirm.append((xmin, ymin, xmax, ymax))
+                    print(f"Bbox revised {t_bboxes} -------> {confirm}")
+            
+            if confirm and visual:
+                # print(filename)
+                visualize(t_image, confirm, labels)
 
+            if confirm:
                 cv2.imwrite(f"{SAVE_DIR}/images/{target}_GAP{GAP}-{idx:>05}.jpg", t_image)
                 write_xml(f"{SAVE_DIR}/annotations", confirm, labels, f"{target}_GAP{GAP}-{idx:>05}", height, width, format='pascal_voc')
-                
-                if confirm and visual:
-                    # print(filename)
-                    visualize(t_image, confirm, labels)
-
-            except:
-                pass
+            else:
+                cv2.imwrite(f"{SAVE_DIR}/images/{target}_{idx:>05}.jpg", t_image)
+                write_xml(f"{SAVE_DIR}/annotations", t_bboxes, labels, f"{target}_{idx:>05}", height, width, format='pascal_voc')
 
     print(label_check_list)
